ws/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
cssutils==2.3.0
DateTime==4.3
Flask==2.0.1
Flask-Cors==3.0.10
fpdf2==2.4.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
pdfdocument==4.0.0
Pillow==8.3.1
premailer==3.10.0
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
reportlab==3.6.1
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
yagmail==0.14.256
zope.interface==5.4.0
